+++ tests/bootstrap.php
@@ -1,0 +1,14 @@
+<?php 
+/**
+ * Library for urls manipulation.
+ * 
+ * @author     Josantonius - hello@josantonius.com
+ * @copyright  Copyright (c) 2017
+ * @license    https://opensource.org/licenses/MIT - The MIT License (MIT)
+ * @link       https://github.com/Josantonius/PHP-Url
+ * @since      1.1.5
+ */
+
+$loader = require __DIR__ . '/../src/bootstrap.php';
+
+$loader->add('Josantonius\Url\Test', __DIR__);